+++ src/client/components/filter/TagsSuggester.component.js
@@ -95,7 +95,7 @@ class TagsSuggester extends React.Component {
 
   render() {
     const inputVisibel = this.state.inputVisibel;
-    const tagsInField = this.props.selectedFilters.length === 0 ? false : true;
+    const tagsInField = this.props.tags.length > 0;
     const inputVisibelClass = inputVisibel || !tagsInField ? '' : '';
     const suggestionlistClass =
       inputVisibel || !tagsInField ? 'input-visible' : 'input-hidden';
@@ -144,6 +144,7 @@ class TagsSuggester extends React.Component {
             onSuggestionSelected={this.props.onSuggestionSelected}
             focusInputOnSuggestionClick={true}
             inputProps={inputProps}
+            highlightFirstSuggestion={true}
           />
         </div>
         {!tagsInField && (
